pyproject.toml
[project]
name = "peyote-pattern"
version = "0.2.0"
description = "Peyote bead pattern generator with text, decorative patterns, and GUI"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.9.0",
    "nicegui>=3.10.0",
    "pillow>=12.2.0",
]

[project.scripts]
peyote = "peyote.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.coverage.run]
source = ["peyote"]
omit = ["peyote/gui.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
